Refactor to add version check to pulumi new as well
